v2607011716_WIP_단면뷰 수정

+++ src/DHBIMWATER.Infrastructure/Repositories/Revit/Modeling/RevitExteriorWallClassifierRepo.cs
@@ -14,7 +14,7 @@ namespace DHBIMWATER.Infrastructure.Repositories.Revit.Modeling
             _doc = doc;
         }
 
-        public IReadOnlyList<(int ElementId, Point2D Start, Point2D End)> GetWallEndpoints()
+        public IReadOnlyList<(long ElementId, Point2D Start, Point2D End)> GetWallEndpoints()
         {
             var doc = _doc();
             if (doc == null) return [];
@@ -26,10 +26,10 @@ namespace DHBIMWATER.Infrastructure.Repositories.Revit.Modeling
                 .Select(wall =>
                 {
                     var curve = (wall.Location as LocationCurve)?.Curve;
-                    if (curve == null) return ((int ElementId, Point2D Start, Point2D End)?)null;
+                    if (curve == null) return ((long ElementId, Point2D Start, Point2D End)?)null;
                     var s = curve.GetEndPoint(0);
                     var e = curve.GetEndPoint(1);
-                    return ((int)wall.Id.Value,
+                    return (wall.Id.Value,
                             new Point2D(UC.FtToMm(s.X), UC.FtToMm(s.Y)),
                             new Point2D(UC.FtToMm(e.X), UC.FtToMm(e.Y)));
                 })
@@ -38,12 +38,12 @@ namespace DHBIMWATER.Infrastructure.Repositories.Revit.Modeling
                 .ToList();
         }
 
-        public void SetExteriorFlag(int elementId, bool isExterior)
+        public void SetExteriorFlag(long elementId, bool isExterior)
         {
             var doc = _doc();
             if (doc == null) return;
 
-            var wall = doc.GetElement(new ElementId((long)elementId)) as Wall;
+            var wall = doc.GetElement(new ElementId(elementId)) as Wall;
             wall?.LookupParameter("DH_IsExterior")?.Set(isExterior ? 1 : 0);
         }
     }
